Исправлен тип данных номера счета с double на int

+++ IntroOOP/BankAccount.cs
@@ -1,11 +1,9 @@
-using System.Xml.Linq;
-
 public class BankAccount
 {
     /// <summary>
     /// Приватнные полы
     /// </summary>
-    private double _Num; // Номер счета    
+    private int _Num; // Номер счета    
     private decimal _Balance; // Баланс
     private Card _TypeCard; // Тип банковского счета
 
@@ -19,7 +17,7 @@ public class BankAccount
     /// <summary>
     /// Чтение и запись приватных полей
     /// </summary>
-    public double Num { get => _Num; set => _Num = value; }
+    public int Num { get => _Num; set => _Num = value; }
     public decimal Balance { get => _Balance; set => _Balance = value; }
     public Card TypeCard { get => _TypeCard; set => _TypeCard = value; }
 
@@ -29,7 +27,7 @@ public class BankAccount
     /// <param name="num">Номер счета</param>
     /// <param name="balance">Баланс</param>
     /// <param name="typeCard">Тип банковского счета</param>
-    public BankAccount(double num, decimal balance, Card typeCard)
+    public BankAccount(int num, decimal balance, Card typeCard)
     {
         _Num = num;
         _Balance = balance;
